ref:整合包安装重写
